DimensionObject refactor (version like 5 now)
- Completely reworked from the ground up many aspects of the DimensionObject system.
-- Now have a DimensionObjectManager that handles rendering and collisions by taking into account ALL DimensionObjects affecting a Renderer/Collider
-- Pillar dimensions are stored internally as a floating point number now instead of integer base dimension + a quadrant. This is a closer representation of a 4D environment where the 4th dimension is the # of times the player has wrapped around the pillar
-- All DimensionObject shaders now use a 256-bit bitmask for rendering. Bitmasks are combined at runtime on the CPU and passed to the GPU as 8 floating point numbers, 32 bits each, in an array for efficient use of data transfer
-- DimensionObjects now share cached Materials created at runtime instead of having a unique Material for each Renderer
-- DimensionObjects now use MaterialProperyBlocks to set their dimension object bitmasks
-- PillarDimensionObjects now store a cyclical interval range of Dimensions for each of its visibility states in order to speed up collision logic checks to two float comparisons
- Fixed bug that caused camera shake to affect the visibility state of PillarDimensionObjects when player was standing on the border
- Some tooling improvements
- Moved UI code to a separate namespace for organization

I might want to plan an IGDA talk about all the different ways I've implemented DimensionObjects, especially PillarDimensionObjects, over the years...

+++ Assets/Editor/Tests/BooleanExpressionTests.cs
@@ -10,7 +10,7 @@ public class BooleanExpressionTests {
         List<string> infix = new List<string>() { "(","0","||","1",")","&&","(","2","&&","(","0","||","3",")",")" };
         List<string> postfix = DimensionObject.InfixToPostfix(infix);
 
-        float[] solutions = DimensionObject.SolutionArrayFromPostfixExpression(postfix);
+        int[] solutions = DimensionObject.SolutionArrayFromPostfixExpression(postfix);
         
         Debug.Log(string.Join(", ", solutions));
     }
